refactor: implement self-hosted px service
